requirements.txt
openai
tavily-python
pandas
python-dotenv
openpyxl
rich
